pyproject.toml
[project]
name = "lightbulbflow"
version = "0.1.0"
description = ""
authors = [
    { name = "Jakub Koláček", email = "jakub@kolacek.dev" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "asyncio == 4.0.0",
    "asyncpraw == 7.8.1",
    "pyyaml == 6.0.3",
    "pydantic == 2.11.10",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
